Fix OSD D-Bus assertion with Qt 6

+++ src/osd/osddbus.cpp
@@ -83,7 +83,7 @@ QDBusArgument &operator<<(QDBusArgument &arg, const QImage &image) {
   arg.beginStructure();
   arg << i.width();
   arg << i.height();
-  arg << i.bytesPerLine();
+  arg << static_cast<qint32>(i.bytesPerLine());
   arg << i.hasAlphaChannel();
   int channels = i.isGrayscale() ? 1 : (i.hasAlphaChannel() ? 4 : 3);
   arg << i.depth() / channels;
